requirements.txt
keras==2.4.3
tensorflow==2.4.1
keras-video-generators
flask
flask-cors
numpy
matplotlib
opencv-python